fix building err

+++ lab2/kern/pmap.c
@@ -104,7 +104,7 @@ boot_alloc(uint32_t n)
 	// LAB 2: Your code here.
 	result = nextfree;
 	nextfree = ROUNDUP(nextfree + n, PGSIZE);
-	if (nextfree - KERNBASE > npages * PGSIZE)
+	if ((uint32_t)nextfree - KERNBASE > npages * PGSIZE)
 		panic("memory out of limit\n");
 
 	return result;
